Cập nhật database để giải quyết các vấn đề như payment audit, oversell

+++ backend/src/main/java/com/deepthocks/backend/dto/ProductCreateDTO.java
@@ -3,6 +3,7 @@ package com.deepthocks.backend.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.*;
@@ -17,7 +18,7 @@ public class ProductCreateDTO {
     private String productDescription;
     private int categoryId;
     private int stockQuantity;
-    private double basePrice;
-    private double salePrice;
+    private BigDecimal basePrice;
+    private BigDecimal salePrice;
     private List<ProductImageDTO> images;
 }
